refactor(network): Use expect/actual for HttpClientEngine
This commit refactors the core-network module to use Kotlin Multiplatform's expect/actual mechanism for providing the HttpClientEngine, replacing the previous Koin-based injection.
An expect function, provideHttpClientEngine(), is now declared in commonMain. Platform-specific actual implementations are provided in androidMain and iosMain to supply the appropriate Ktor engine for each platform.
This change simplifies the dependency setup for the network client and better aligns the module with KMP best practices for handling platform-specific implementations.
