+++ androidApp/build.gradle.kts
@@ -1,4 +1,3 @@
-
 plugins {
     alias(libs.plugins.androidApplication)
     alias(libs.plugins.moviedb.composeMultiplatform)
@@ -6,7 +5,6 @@ plugins {
 
 dependencies {
     implementation(projects.composeApp)
-    implementation(projects.core.network)
     implementation(projects.core.ui)
 
     implementation(libs.koin.android)
