+++ androidApp/src/main/kotlin/com/elna/moviedb/MovieDBApp.kt
@@ -1,7 +1,6 @@
 package com.elna.moviedb
 
 import android.app.Application
-import com.elna.moviedb.core.network.di.androidDataModule
 import com.elna.moviedb.di.initKoin
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
@@ -15,7 +14,6 @@ class MovieDBApp : Application() {
             appDeclaration = {
                 androidLogger()
                 androidContext(this@MovieDBApp)
-                modules(androidDataModule)
             }
         )
     }
