--- composeApp/src/iosMain/kotlin/com/elna/moviedb/di/Helper.kt
@@ -1,9 +1,0 @@
-package com.elna.moviedb.di
-
-fun initKoin() {
-    initKoin(
-        appDeclaration = {
-            modules(iOSModule)
-        }
-    )
-}
