--- composeApp/src/iosMain/kotlin/com/elna/moviedb/di/IosModule.kt
@@ -1,9 +1,0 @@
-package com.elna.moviedb.di
-
-
-import io.ktor.client.engine.darwin.Darwin
-import org.koin.dsl.module
-
-val iOSModule = module {
-    single { Darwin.create() } //for inject httpClientEngine
-}
