+++ core/network/src/androidMain/kotlin/com/elna/moviedb/core/network/client/HttpClientEngineProvider.android.kt
@@ -1,0 +1,8 @@
+package com.elna.moviedb.core.network.client
+
+import io.ktor.client.engine.HttpClientEngineFactory
+import io.ktor.client.engine.android.Android
+
+actual fun provideHttpClientEngine(): HttpClientEngineFactory<*> {
+    return Android
+}
