--- core/network/src/androidMain/kotlin/com/elna/moviedb/core/network/di/AndroidModule.kt
@@ -1,12 +1,0 @@
-package com.elna.moviedb.core.network.di
-
-
-import io.ktor.client.engine.android.Android
-import org.koin.dsl.module
-
-/**
- * Android platform specific module
- */
-val androidDataModule = module {
-    single { Android.create() }  //for inject httpClientEngine
-}
