+++ core/network/src/commonMain/kotlin/com/elna/moviedb/core/network/client/HttpClientConfig.kt
@@ -1,7 +1,7 @@
-package com.elna.moviedb.core.network.ktor
+package com.elna.moviedb.core.network.client
 
 import io.ktor.client.HttpClient
-import io.ktor.client.engine.HttpClientEngine
+import io.ktor.client.engine.HttpClientEngineFactory
 import io.ktor.client.plugins.HttpTimeout
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
 import io.ktor.client.plugins.logging.LogLevel
@@ -10,7 +10,7 @@ import io.ktor.client.plugins.logging.Logging
 import io.ktor.serialization.kotlinx.json.json
 import kotlinx.serialization.json.Json
 
-fun createHttpClient(httpClientEngine: HttpClientEngine) = HttpClient(httpClientEngine) {
+fun createHttpClient(engine: HttpClientEngineFactory<*>) = HttpClient(engine) {
     install(ContentNegotiation) {
         json(Json {
             prettyPrint = true
