+++ core/network/src/commonMain/kotlin/com/elna/moviedb/core/network/client/HttpClientEngineProvider.kt
@@ -1,0 +1,5 @@
+package com.elna.moviedb.core.network.client
+
+import io.ktor.client.engine.HttpClientEngineFactory
+
+expect fun provideHttpClientEngine(): HttpClientEngineFactory<*>
