+++ core/network/src/iosMain/kotlin/com/elna/moviedb/core/network/client/HttpClientEngineProvider.ios.kt
@@ -1,0 +1,8 @@
+package com.elna.moviedb.core.network.client
+
+import io.ktor.client.engine.HttpClientEngineFactory
+import io.ktor.client.engine.darwin.Darwin
+
+actual fun provideHttpClientEngine(): HttpClientEngineFactory<*> {
+    return Darwin
+}
